move modules to common dir
