+++ tensorrt_llm/_torch/visual_gen/models/wan/__init__.py
@@ -1,7 +1,7 @@
+from .parallel_vae import WanParallelVAEAdapter
 from .pipeline_wan import WanPipeline
 from .pipeline_wan_i2v import WanImageToVideoPipeline
 from .transformer_wan import WanTransformer3DModel
-from .vae import WanParallelVAEAdapter
 
 __all__ = [
     "WanPipeline",
