+++ tensorrt_llm/_torch/visual_gen/models/wan/pipeline_wan.py
@@ -16,8 +16,8 @@ from tensorrt_llm._torch.visual_gen.utils import postprocess_video_tensor
 from tensorrt_llm._utils import nvtx_range
 from tensorrt_llm.logger import logger
 
+from .parallel_vae import WanParallelVAEAdapter
 from .transformer_wan import WanTransformer3DModel
-from .vae import WanParallelVAEAdapter
 
 # Supported Wan T2V models:
 # - Wan2.1-T2V-14B: Single-stage text-to-video (14B parameters)
