+++ tensorrt_llm/_torch/visual_gen/modules/__init__.py
@@ -19,15 +19,8 @@ This module provides modular neural network components for visual generation mod
 """
 
 from .attention import Attention, QKVMode
-from .conv import HaloExchangeConv, HaloExchangeConv2dStride2
-from .norm import GroupNormParallel
-from .parallel_vae import BaseParallelVAEAdapter
 
 __all__ = [
     "Attention",
     "QKVMode",
-    "HaloExchangeConv",
-    "HaloExchangeConv2dStride2",
-    "GroupNormParallel",
-    "BaseParallelVAEAdapter",
 ]
