+++ tensorrt_llm/_torch/visual_gen/modules/vae/__init__.py
@@ -1,0 +1,12 @@
+from .attention import ParallelVaeAttentionBlock
+from .conv import HaloExchangeConv, HaloExchangeConv2dStride2
+from .norm import GroupNormParallel
+from .parallel_vae_interface import BaseParallelVAEAdapter
+
+__all__ = [
+    "ParallelVaeAttentionBlock",
+    "HaloExchangeConv",
+    "HaloExchangeConv2dStride2",
+    "GroupNormParallel",
+    "BaseParallelVAEAdapter",
+]
